+++ src/pages/programsWip/form/EnrollmentDataFormContents.tsx
@@ -11,7 +11,8 @@ import {
     TransferOption,
 } from '@dhis2/ui'
 import React, { useEffect, useRef } from 'react'
-import { Field as FieldRFF, useField } from 'react-final-form'
+import { useQuery } from '@tanstack/react-query'
+import { Field as FieldRFF, FieldRenderProps, useField } from 'react-final-form'
 import {
     ModelTransfer,
     RenderingOptionsSelect,
@@ -25,8 +26,14 @@ import {
     InfoIconWithTooltip,
     TooltipWrapper,
 } from '../../../components/tooltip'
+import { getConstantTranslation, useBoundResourceQueryFn } from '../../../lib'
 import { ProgramTrackedEntityAttribute } from '../../../types/generated'
 import { ProgramsFromFilters } from '../Edit'
+
+const defaultRenderType = {
+    MOBILE: { type: 'DEFAULT' },
+    DESKTOP: { type: 'DEFAULT' },
+}
 
 const defaultRenderType = {
     MOBILE: { type: 'DEFAULT' },
@@ -151,7 +158,10 @@ export const EnrollmentDataFormContents = React.memo(
                         <ModelTransfer
                             selected={input.value.map((attribute) => {
                                 const tea = attribute.trackedEntityAttribute
-                                return tea
+                                return {
+                                    ...tea,
+                                    disabled: tetaIds.has(tea.id),
+                                }
                             })}
                             renderOption={({ value, ...rest }) => {
                                 const tea = value
@@ -186,27 +196,6 @@ export const EnrollmentDataFormContents = React.memo(
                                     ])
                                 )
 
-                                const tetaOriginalIndices = new Map<
-                                    string,
-                                    number
-                                >()
-                                input.value.forEach((attr, index) => {
-                                    if (
-                                        tetaIds.has(
-                                            attr.trackedEntityAttribute.id
-                                        )
-                                    ) {
-                                        tetaOriginalIndices.set(
-                                            attr.trackedEntityAttribute.id,
-                                            index
-                                        )
-                                    }
-                                })
-
-                                const selectedIds = new Set(
-                                    selected.map((s) => s.id)
-                                )
-
                                 const selectedAttributes = selected.map((s) => {
                                     const existing = existingAttributesMap.get(
                                         s.id
@@ -230,42 +219,7 @@ export const EnrollmentDataFormContents = React.memo(
                                     }
                                 })
 
-                                // Temperary functionoutality to re-insert TETAs that the user tried to remove
-                                // Should be removed if transfer arrow buttons gets conditionally disabling
-                                const missingTetas = input.value
-                                    .filter(
-                                        (attr) =>
-                                            tetaIds.has(
-                                                attr.trackedEntityAttribute.id
-                                            ) &&
-                                            !selectedIds.has(
-                                                attr.trackedEntityAttribute.id
-                                            )
-                                    )
-                                    .map((teta) => ({
-                                        teta: {
-                                            ...teta,
-                                            renderType:
-                                                teta.renderType ||
-                                                defaultRenderType,
-                                        },
-                                        index:
-                                            tetaOriginalIndices.get(
-                                                teta.trackedEntityAttribute.id
-                                            ) ?? Infinity,
-                                    }))
-                                    .sort((a, b) => a.index - b.index)
-
-                                const result = [...selectedAttributes]
-                                missingTetas.forEach(({ teta, index }) => {
-                                    result.splice(
-                                        Math.min(index, result.length),
-                                        0,
-                                        teta
-                                    )
-                                })
-
-                                input.onChange(result)
+                                input.onChange(selectedAttributes)
                                 input.onBlur()
                             }}
                             leftHeader={i18n.t('Available attributes')}