+++ src/pages/programsWip/form/RoleAccess/index.ts
@@ -1,0 +1,2 @@
+export { RoleAccess } from './RoleAccess'
+export { RoleAccessBox } from './RoleAccessBox'